# Speed up globbing relative to given directories