--- a/lib/rubygems/installer.rb
+++ b/lib/rubygems/installer.rb
@@ -379,7 +379,7 @@ class Gem::Installer
     @specs ||= begin
       specs = []
 
-      Dir[File.join(gem_home, "specifications", "*.gemspec")].each do |path|
+      Gem::Util.glob_files_in_dir("*.gemspec", File.join(gem_home, "specifications")).each do |path|
         spec = Gem::Specification.load path.untaint
         specs << spec if spec
       end