--- a/lib/rubygems/util.rb
+++ b/lib/rubygems/util.rb
@@ -116,4 +116,16 @@ module Gem::Util
     end
   end
 
+  ##
+  # Globs for files matching +pattern+ inside of +directory+,
+  # returning absolute paths to the matching files.
+
+  def self.glob_files_in_dir(glob, base_path)
+    if RUBY_VERSION >= "2.5"
+      Dir.glob(glob, base: base_path).map! {|f| File.join(base_path, f) }
+    else
+      Dir.glob(File.expand_path(glob, base_path))
+    end
+  end
+
 end