# Description:

This is picking up @tenderlove's improvements in https://github.com/rubygems/rubygems/pull/971.

Benchmark:

```ruby
class Hello
  def self.old_find_files_from_load_path glob # :nodoc:
    $LOAD_PATH.map { |load_path|
      Dir["#{File.expand_path glob, load_path}#{Gem.suffix_pattern}"]
    }.flatten.select { |file| File.file? file.untaint }
  end

  def self.aaron_find_files_from_load_path glob # :nodoc:
    search = "#{glob}#{Gem.suffix_pattern}"
    $LOAD_PATH.map { |load_path|
      Dir.chdir(load_path) {
        Dir[search].map! { |f| File.join load_path, f }
      }
    }.flatten.select { |file| File.file? file.untaint }
  end
  
  def self.new_find_files_from_load_path glob # :nodoc:
    glob_with_suffixes = "#{glob}#{Gem.suffix_pattern}"
    $LOAD_PATH.map { |load_path|
      Dir.glob(glob_with_suffixes, base: load_path).map! {|f| File.join(load_path, f) }
    }.flatten.select { |file| File.file? file.untaint }
  end
end

search = "minitest/*_plugin.rb"

Benchmark.ips do |x|
  x.report('old') do
    Hello.old_find_files_from_load_path search
  end

  x.report('aaron') do
    Hello.aaron_find_files_from_load_path search
  end
  
  x.report('new') do
    Hello.new_find_files_from_load_path search
  end
  
  x.compare!
end
```

Results (on Ruby 2.5):

```
Warming up --------------------------------------
                 old    94.000  i/100ms
               aaron   297.000  i/100ms
                 new   729.000  i/100ms
Calculating -------------------------------------
                 old    840.927  (±28.2%) i/s -      3.666k in   5.094465s
               aaron      2.843k (±19.2%) i/s -     13.365k in   5.012555s
                 new      7.457k (± 9.4%) i/s -     37.179k in   5.046804s

Comparison:
                 new:     7457.4 i/s
               aaron:     2842.6 i/s - 2.62x  slower
                 old:      840.9 i/s - 8.87x  slower
```

# Tasks:

- [ ] Describe the problem / feature
- [ ] Write tests
- [ ] Write code to solve the problem
- [ ] Get code review from coworkers / friends

I will abide by the [code of conduct](https://github.com/rubygems/rubygems/blob/master/CODE_OF_CONDUCT.md).