@MSP-Greg 2018-06-28
Is it a fast system with a fast SSD, or an older system with a hard drive?  How many files searched, how many files match?  Etc...